+++ bindings/matrix-sdk-crypto-js/CHANGELOG.md
@@ -1,6 +1,6 @@
 # v0.1.0-alpha.11
 
--   Add `VerificationRequest.phase()`
+-   Add `VerificationRequest.phase()` and `VerificationRequest.getVerification()`.
 
 # v0.1.0-alpha.10